+++ assignment/src/main/java/com/davidholas/assignment/model/TransferDetails.java
@@ -1,14 +1,16 @@
 package com.davidholas.assignment.model;
 
+import java.math.BigDecimal;
+
 public class TransferDetails {
 
     private Long withdrawalAccountId;
 
     private Long depositAccountId;
 
-    private double amount;
+    private BigDecimal amount;
 
-    public TransferDetails(Long withdrawalAccountId, Long depositAccountId, double amount) {
+    public TransferDetails(Long withdrawalAccountId, Long depositAccountId, BigDecimal amount) {
         this.withdrawalAccountId = withdrawalAccountId;
         this.depositAccountId = depositAccountId;
         this.amount = amount;
@@ -30,11 +32,11 @@ public class TransferDetails {
         this.depositAccountId = depositAccountId;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
